attachment_sandbox_service/requirements.txt
fastapi>=0.115,<1
uvicorn>=0.30,<1
python-multipart>=0.0.9,<1
httpx>=0.27,<1
pydantic-settings>=2.6,<3
sqlalchemy>=2.0,<3
greenlet>=3.1,<4
aiosqlite>=0.20,<1
asyncpg>=0.29,<1
redis>=5.0,<6
yara-python>=4.5,<5
olefile>=0.47,<1
oletools>=0.60,<1
pefile>=2024.8.26,<2026
LnkParse3>=1.5.3,<2
pdfminer.six>=20231228,<20270000
pypdf>=5.0,<6
pytest>=7.4,<8
